[deployment] ProcessMessage explicitly indicates timer reset
The complication here is that when we are in a waiting state, we don't
want converge requests to reset our timer since the "nextState" is the
same as the previous state.

This commit changes ProcessMessage to explicitly signal whether the
timer needs to be reset. Alternatively, ProcessMessage could be
responsible for calling Stop() on any timers before returning.

Signed-off-by: Steven Danna <steve@chef.io>